Add Kubernetes mixin
This change adds alerting rules, recording rules and dashboard, which
are auto-generated based on the Kubernetes mixin. Because we do not
deploy our own cadvisor, I changed the OpenShift Prometheus
configuration to federate all platform metrics. If this turns out to be
too high of a volume, we can restrict this again and only federate
metrics required by the mixin.

This adds about ~60 new alerting rules, which may create some initial
noise. Unwanted alerts can be silenced in the Alertmanager and then
disabled in the mixin configuration.