always extend SLO calculation over entire 28 days
Extended average over time refers to the time series effectively being extended
over the entire time interval. This is in contrast to avg_over_time, which
only averages over time intervals where the time series is not nil.
This is important during the initial 28 days of Central instances. For example,
consider a Central instance that lived for 5 minutes and was down for 2 minutes.
Using avg_over_time, the availability would be 3 min / 5 min = 60%. The
extended average over 28 days would yield 1 - 2 min / 28 days ~ 99.995%.
After the initial 28 days, both averages are equivalent.